image_description_app/gradio_service/requirements.txt
gradio
aiohttp
python-dotenv
Pillow